+++ SmoothCamera_Engine/AppVersion.cs
@@ -6,7 +6,7 @@ namespace SmoothCamera_Engine
         public const string VersionChannel = "PR-R";
         public const int CycleBuildNumber = 2;
         public const string ChangeLetters = "SP";
-        public const int SubNumber = 90;
+        public const int SubNumber = 91;
 
         public static string BuildToken => $"{VersionChannel}{CycleBuildNumber}{ChangeLetters}{SubNumber}";
         public static string Display => $"{ReleaseBase} Build {BuildToken}";